Split JSON API utils to support multiple ORM
