+++ lib/configuration/hooks/jsonapi/helpers/response.js
@@ -40,7 +40,7 @@ module.exports = {
     }
 
     // Fetch and format value
-    const value = this.fetchValue(ctx, object);
+    const value = this.fetchValue(ctx, object).toJSON() || this.fetchValue(ctx, object);
 
     if (!_.isEmpty(value)) {
       ctx.response.body = this.serialize(ctx, type, object, value, matchedRoute);
@@ -54,7 +54,7 @@ module.exports = {
   serialize: function (ctx, type, object, value, matchedRoute) {
     const toSerialize = {
       topLevelLinks: {self: ctx.request.origin + ctx.request.url},
-      keyForAttribute: 'camelCase',
+      keyForAttribute: 'dash-case',
       pluralizeType: false,
       included: true,
       typeForAttribute: function (currentType) {
@@ -129,6 +129,9 @@ module.exports = {
     if (_.isPlainObject(strapi.config.jsonapi) && strapi.config.jsonapi.hasOwnProperty('pagination') && strapi.config.jsonapi.pagination === true) {
       this.includePagination(ctx, toSerialize, object, type, matchedRoute);
     }
+
+    console.log(toSerialize);
+    console.log(value);
 
     const serialized = new JSONAPI.Serializer(type, value, toSerialize);
