+++ lib/configuration/hooks/jsonapi/utils/utils.js
@@ -6,6 +6,8 @@
 
 // Public node modules.
 const _ = require('lodash');
+const utilsBookShelf = require('./bookshelf');
+const utilsWaterline = require('./waterline');
 
 /**
  * JSON API utils
@@ -85,17 +87,14 @@ module.exports = {
       return null;
     }
 
-    const PK = _.findKey(strapi.models[type].attributes, {primaryKey: true});
-
-    if (!_.isUndefined(PK)) {
-      return PK;
-    } else if (strapi.models[type].attributes.hasOwnProperty('id')) {
-      return 'id';
-    } else if (strapi.models[type].attributes.hasOwnProperty('uuid')) {
-      return 'uuid';
+    switch (strapi.models[type].orm.toLowerCase()) {
+      case 'bookshelf':
+        return utilsBookShelf.getPK(type);
+      case 'waterline':
+        return utilsWaterline.getPK(type);
+      default:
+        return null;
     }
-
-    return null;
   },
 
   /**
